pyproject.toml
[tool.poetry]
name = "ml_prediction_web_service"
version = "0.0.2"
authors = ["Arman Hovhannisyan <hovhannisyanarman062@gmail.com>"]
description = "Perovskite analysis with ML"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.0"
pandas = "^2.2.3"
numpy = "1.26.4"
matplotlib = "3.10.0"
plotly = "6.4.0"
seaborn = "0.13.2"
scikit-learn = "1.6.1"
scipy = "^1.7.3"
pyarrow = "^18.1.0"
httpx = "^0.28.1"
google-api-python-client = "2.163.0"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.1"
openpyxl = "3.1.5"
xgboost = "2.1.4"
catboost = "1.2.8"
fastapi = "0.115.12"
sqlalchemy = "2.0.44"
uvicorn = "0.37.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.23.2"
pytest-dotenv = "^0.5.2"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

pythonpath = [
    "src",
    "tests"
]
